fix: move from SharedComponents to Components
